+++ Scripts/create_forecast.sh
@@ -4,7 +4,7 @@
 
 if [ $# -ne 3 ]
 then
-  echo "Usage: $0 enssize date_deb date_obs"
+  echo "Usage: $0 date_deb enssize date_obs"
   exit 1
 else
   date=$1
@@ -14,7 +14,7 @@ fi
 
 cd ${ROOT_DIR}/ASSIM
 
-\rm forecast???.nc
+[ -f forecast001.nc ] && \rm forecast???.nc
 
 let Nensp1=enssize+1
 
@@ -27,14 +27,14 @@ do
    if [ $SMOOTH == 0 ]
    then 
      # link to the current state
-     ln -s ${FORDIR}/${CASEDIR}${count2}/shoreface_out.nc forecast${count2}.nc
+     ln -fs ${FORDIR}/${CASEDIR}${count2}/shoreface_out.nc forecast${count2}.nc
    else
      # smoother
      #solution at time_obs
      ln -s ${FORDIR}/${CASEDIR}${count2}/ARCHIVE/shoreface_trajectory${count2}_${dateobs}.nc trajectory${count2}_T001.nc
      
      #solution atbeginning time
-     cp ${FORDIR}/${CASEDIR}${count2}/ARCHIVE/shoreface_forecast${count2}_${date}.nc ${FORDIR}/${CASEDIR}${count2}/shoreface_out.nc
+     \cp ${FORDIR}/${CASEDIR}${count2}/ARCHIVE/shoreface_forecast${count2}_${date}.nc ${FORDIR}/${CASEDIR}${count2}/shoreface_out.nc
      ln -s ${FORDIR}/${CASEDIR}${count2}/shoreface_out.nc forecast${count2}.nc
    fi
